move project answers endpoints to its own file
